+++ Model/Timeline/Data/DbWaypoint.cs
@@ -8,8 +8,8 @@ namespace Model.Timeline.Data
     {
         [Key]
         public int Id { get; set; }
-        public int LatitudeE7 { get; set; }
-        public int LongitudeE7 { get; set; }
+        public long LatitudeE7 { get; set; }
+        public long LongitudeE7 { get; set; }
         public int? DbActivitySegmentId { get; set; }
 
         public DbWaypoint()
